+++ js/media/view.js
@@ -39,10 +39,9 @@
 		},
 
 		render: function() {
-			var self, model, chart, gv, type, series, data, table, settings, i, j, row, date, format, formatter, axis, property;
+			var self, model, chart;
 
 			self = this;
-			gv = google.visualization;
 			model = self.model;
 
 			self.$el
@@ -50,137 +49,17 @@
 				.height(self.options.height)
 				.css('background-image', 'none');
 
-			type = model.get('type');
-			series = model.get('series');
-			data = model.get('data');
-			settings = model.get('settings');
+            chart = {};
+			chart.type = model.get('type');
+			chart.series = model.get('series');
+			chart.data = model.get('data');
+			chart.library = model.get('library');
+			chart.settings = model.get('settings');
+			chart.settings.width = self.options.width;
+			chart.settings.height = self.options.height;
+            $('#' + self.id).append(model.get('css'));
 
-			settings.width = self.options.width;
-			settings.height = self.options.height;
-
-			table = new gv.DataTable({cols: series});
-
-            switch (type) {
-                case "gauge":
-                case "table":
-                case "timeline":
-                    chart  = type.charAt(0).toUpperCase() + type.slice(1);
-                    break;
-                default:
-			        chart = type.charAt(0).toUpperCase() + type.slice(1) + 'Chart';
-                    break;
-            }
-
-			chart = new gv[chart](self.el);
-
-			switch (type) {
-				case 'pie':
-					if (settings.slices) {
-						for (i in settings.slices) {
-							if (settings.slices[i]['color'] === '') {
-								delete settings.slices[i]['color'];
-							}
-						}
-					}
-					break;
-				case 'line':
-				case 'bar':
-				case 'column':
-				case 'area':
-				case 'scatter':
-				case 'candlestick':
-					if (settings.series) {
-						for (i in settings.series) {
-							if (settings.series[i]['color'] === '') {
-								delete settings.series[i]['color'];
-							}
-						}
-					}
-					break;
-				case 'geo':
-					if (settings.region !== undefined && settings.region.replace(/^\s+|\s+$/g, '') === '') {
-						settings['region'] = 'world';
-					}
-					break;
-				case 'gauge':
-					break;
-				default:
-					return;
-			}
-
-			if (series[0] && (series[0].type === 'date' || series[0].type === 'datetime')) {
-				axis = false;
-				switch (type) {
-					case 'line':
-					case 'area':
-					case 'scatter':
-					case 'candlestick':
-					case 'column':
-						axis = settings.hAxis;
-						break;
-					case 'bar':
-						axis = settings.vAxis;
-						break;
-				}
-
-				if (axis) {
-					for (property in axis.viewWindow) {
-						date = new Date(axis.viewWindow[property]);
-						if (Object.prototype.toString.call(date) === "[object Date]") {
-							if (!isNaN(date.getTime())) {
-								axis.viewWindow[property] = date;
-								continue;
-							}
-						}
-
-						delete axis.viewWindow[property];
-					}
-				}
-			}
-
-			for (i = 0; i < data.length; i++) {
-				row = [];
-				for (j = 0; j < series.length; j++) {
-					if (series[j].type === 'date' || series[j].type === 'datetime') {
-						date = new Date(data[i][j]);
-						data[i][j] = null;
-						if (Object.prototype.toString.call(date) === "[object Date]") {
-							if (!isNaN(date.getTime())) {
-								data[i][j] = date;
-							}
-						}
-					}
-					row.push(data[i][j]);
-				}
-				table.addRow(row);
-			}
-
-			if (settings.series) {
-				for (i = 0; i < settings.series.length; i++) {
-					format = settings.series[i].format;
-					if (!format || format === '') {
-						continue;
-					}
-
-					formatter = null;
-					switch (series[i + 1].type) {
-						case 'number':
-							formatter = new gv.NumberFormat({pattern: format});
-							break;
-						case 'date':
-						case 'datetime':
-						case 'timeofday':
-							formatter = new gv.DateFormat({pattern: format});
-							break;
-					}
-
-					if (formatter) {
-						formatter.format(table, i + 1);
-					}
-				}
-			}
-
-			chart.draw(table, settings);
+            $('body').trigger('visualizer:render:specificchart:start', {id: self.id, chart: chart, v: {page_type: 'post'}} );
 		}
 	});